+++ src/app/views/large_input_editor.h
@@ -7,5 +7,6 @@
 namespace app::views {
 
 void build_large_input_editor(eui::Ui& ui, const ViewContext& context);
+bool apply_large_input_history(bool redo);
 
 } // namespace app::views